Autocomplete: reuse `generateCompletions` across providers
