+++ vscode/src/completions/providers/google.ts
@@ -1,17 +1,6 @@
 import type { CodeCompletionsParams } from '@sourcegraph/cody-shared'
 
-import { forkSignal, generatorWithTimeout, zipGenerators } from '../utils'
-
-import {
-    type FetchCompletionResult,
-    fetchAndProcessDynamicMultilineCompletions,
-} from './shared/fetch-and-process-completions'
-import {
-    type CompletionProviderTracer,
-    type GenerateCompletionsOptions,
-    Provider,
-    type ProviderFactoryParams,
-} from './shared/provider'
+import { type GenerateCompletionsOptions, Provider, type ProviderFactoryParams } from './shared/provider'
 
 class GoogleGeminiProvider extends Provider {
     public getRequestParams(options: GenerateCompletionsOptions): CodeCompletionsParams {
@@ -31,39 +20,6 @@ class GoogleGeminiProvider extends Provider {
             model: `${this.id}/${this.legacyModel}`,
             messages,
         }
-    }
-
-    public async generateCompletions(
-        generateOptions: GenerateCompletionsOptions,
-        abortSignal: AbortSignal,
-        tracer?: CompletionProviderTracer
-    ): Promise<AsyncGenerator<FetchCompletionResult[]>> {
-        const { numberOfCompletionsToGenerate, docContext } = generateOptions
-
-        const requestParams = this.getRequestParams(generateOptions)
-        tracer?.params(requestParams)
-
-        const completionsGenerators = Array.from({ length: numberOfCompletionsToGenerate }).map(
-            async () => {
-                const abortController = forkSignal(abortSignal)
-
-                const completionResponseGenerator = generatorWithTimeout(
-                    await this.client.complete(requestParams, abortController),
-                    requestParams.timeoutMs,
-                    abortController
-                )
-
-                return fetchAndProcessDynamicMultilineCompletions({
-                    completionResponseGenerator,
-                    abortController,
-                    generateOptions,
-                    providerSpecificPostProcess: content =>
-                        this.modelHelper.postProcess(content, docContext),
-                })
-            }
-        )
-
-        return zipGenerators(await Promise.all(completionsGenerators))
     }
 }
