+++ vscode/src/completions/providers/openaicompatible.ts
@@ -1,18 +1,8 @@
 import { type CodeCompletionsParams, charsToTokens, logError } from '@sourcegraph/cody-shared'
 
-import { forkSignal, generatorWithTimeout, zipGenerators } from '../utils'
-
 import { logDebug } from '../../log'
-import {
-    type FetchCompletionResult,
-    fetchAndProcessDynamicMultilineCompletions,
-} from './shared/fetch-and-process-completions'
-import {
-    type CompletionProviderTracer,
-    type GenerateCompletionsOptions,
-    Provider,
-    type ProviderFactoryParams,
-} from './shared/provider'
+
+import { type GenerateCompletionsOptions, Provider, type ProviderFactoryParams } from './shared/provider'
 
 class OpenAICompatibleProvider extends Provider {
     public getRequestParams(options: GenerateCompletionsOptions): CodeCompletionsParams {
@@ -31,54 +21,6 @@ class OpenAICompatibleProvider extends Provider {
             ...this.defaultRequestParams,
             messages,
         }
-    }
-
-    public async generateCompletions(
-        generateOptions: GenerateCompletionsOptions,
-        abortSignal: AbortSignal,
-        tracer?: CompletionProviderTracer
-    ): Promise<AsyncGenerator<FetchCompletionResult[]>> {
-        const { docContext, numberOfCompletionsToGenerate } = generateOptions
-
-        const requestParams = this.getRequestParams(generateOptions)
-        tracer?.params(requestParams)
-
-        const completionsGenerators = Array.from({ length: numberOfCompletionsToGenerate }).map(
-            async () => {
-                const abortController = forkSignal(abortSignal)
-
-                const completionResponseGenerator = generatorWithTimeout(
-                    await this.client.complete(requestParams, abortController),
-                    requestParams.timeoutMs,
-                    abortController
-                )
-
-                return fetchAndProcessDynamicMultilineCompletions({
-                    completionResponseGenerator,
-                    abortController,
-                    generateOptions,
-                    providerSpecificPostProcess: content =>
-                        this.modelHelper.postProcess(content, docContext),
-                })
-            }
-        )
-
-        /**
-         * This implementation waits for all generators to yield values
-         * before passing them to the consumer (request-manager). While this may appear
-         * as a performance bottleneck, it's necessary for the current design.
-         *
-         * The consumer operates on promises, allowing only a single resolve call
-         * from `requestManager.request`. Therefore, we must wait for the initial
-         * batch of completions before returning them collectively, ensuring all
-         * are included as suggested completions.
-         *
-         * To circumvent this performance issue, a method for adding completions to
-         * the existing suggestion list is needed. Presently, this feature is not
-         * available, and the switch to async generators maintains the same behavior
-         * as with promises.
-         */
-        return zipGenerators(await Promise.all(completionsGenerators))
     }
 }
